requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.5
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
django_debug_toolbar==3.8.1
kombu==5.2.4
Pillow==9.5.0
prompt-toolkit==3.0.38
psycopg2==2.9.5
pytz==2023.3
redis==4.5.4
six==1.16.0
sqlparse==0.4.3
vine==5.0.0
wcwidth==0.2.6
